@@ -27,7 +27,7 @@
         }
     ],
     "require": {
-        "php": "^7.4|^8.0",
+        "php": "^8.0",
         "psr/container": "^1.0|^2.0",
         "yiisoft/definitions": "^1.0|^2.0",
         "yiisoft/factory": "^1.0",
@@ -36,6 +36,7 @@
     },
     "require-dev": {
         "phpunit/phpunit": "^9.5",
+        "rector/rector": "^0.14.3",
         "roave/infection-static-analysis-plugin": "^1.16",
         "spatie/phpunit-watcher": "^1.23",
         "vimeo/psalm": "^4.18",
