@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Yiisoft\Widget;
 
 use RuntimeException;
+use Stringable;
 use Yiisoft\Definitions\Exception\CircularReferenceException;
 use Yiisoft\Definitions\Exception\InvalidConfigException;
 use Yiisoft\Definitions\Exception\NotInstantiableException;
@@ -13,7 +14,6 @@ use Yiisoft\Html\NoEncodeStringableInterface;
 
 use function array_key_exists;
 use function array_pop;
-use function get_class;
 use function is_array;
 use function sprintf;
 
@@ -23,7 +23,7 @@ use function sprintf;
  *
  * This is the base class that is meant to be inherited when implementing your own widgets.
  */
-abstract class Widget implements NoEncodeStringableInterface
+abstract class Widget implements NoEncodeStringableInterface, Stringable
 {
     /**
      * The widgets that are currently opened and not yet closed.
@@ -36,7 +36,7 @@ abstract class Widget implements NoEncodeStringableInterface
     /**
      * Used to open a wrapping widget (the one with begin/end).
      *
-     * When implementing this method, don't forget to call parent::begin().
+     * When implementing this method, don't forget to call `parent::begin()`.
      *
      * @return string|null Opening part of widget markup.
      */
@@ -62,7 +62,7 @@ abstract class Widget implements NoEncodeStringableInterface
         }
 
         $widget = array_pop(self::$stack);
-        $widgetClass = get_class($widget);
+        $widgetClass = $widget::class;
 
         if ($widgetClass !== static::class) {
             throw new RuntimeException(sprintf(
@@ -87,7 +87,7 @@ abstract class Widget implements NoEncodeStringableInterface
      *
      * @return static The widget instance.
      */
-    final public static function widget($config = []): self
+    final public static function widget(array|callable|string $config = []): self
     {
         if (is_array($config) && !array_key_exists('class', $config)) {
             $config['class'] = static::class;
